Create a map of dodgr flows.

Create a map of the output of dodgr_flows_aggregate or dodgr_flows_disperse

Usage

dodgr_flowmap(net, bbox = NULL, linescale = 1)

Arguments

net

A street network with a flow column obtained from dodgr_flows_aggregate or dodgr_flows_disperse

bbox

If given, scale the map to this bbox, otherwise use entire extend of net

linescale

Maximal thickness of plotted lines

Note

net should be first passed through merge_directed_graph prior to plotting, otherwise lines for different directions will be overlaid.

See also

Examples

graph <- weight_streetnet (hampi)
from <- sample (graph$from_id, size = 10)
to <- sample (graph$to_id, size = 5)
to <- to [!to %in% from]
flows <- matrix (
    10 * runif (length (from) * length (to)),
    nrow = length (from)
)
graph <- dodgr_flows_aggregate (graph, from = from, to = to, flows = flows)
# graph then has an additonal 'flows` column of aggregate flows along all
# edges. These flows are directed, and can be aggregated to equivalent
# undirected flows on an equivalent undirected graph with:
graph_undir <- merge_directed_graph (graph)
if (FALSE) { # \dontrun{
dodgr_flowmap (graph_undir)
} # }