dodgr_flows

Aggregate flows throughout a network based on an input matrix of flows between all pairs of from and to points.

dodgr_flows(graph, from, to, flows, wt_profile = "bicycle", heap = "BHeap",
  quiet = TRUE)

Arguments

graph
data.frame or equivalent object representing the network graph (see Details)

from
Vector or matrix of points **from** which route distances are to be calculated (see Details)

to
Vector or matrix of points **to** which route distances are to be calculated (see Details)

flows
Matrix of flows with nrow(flows)==length(from) and ncol(flows)==length(to).

wt_profile
Name of weighting profile for street networks (one of foot, horse, wheelchair, bicycle, moped, motorcycle, motorcar, goods, hgv, psv).

heap
Type of heap to use in priority queue. Options include Fibonacci Heap (default; FHeap), Binary Heap (BHeap), Radix, Trinomial Heap (TriHeap), Extended Trinomial Heap (TriHeapExt, and 2-3 Heap (Heap23).

quiet
If FALSE, display progress messages on screen.

Value

Modified version of graph with additonal flow column added.

Examples

graph <- weight_streetnet (hampi)
from <- sample (graph$from_id, size = 10)
to <- sample (graph$to_id, size = 5)
to <- to [!to %in% from]
flows <- matrix (10 * runif (length (from) * length (to)),
                 nrow = length (from))
graph <- dodgr_flows (graph, from = from, to = to, flows = flows)
# graph then has an additonal 'flows` column of aggregate flows along all
# edges