Load a street network previously saved with dodgr_save_streetnet.

This always returns the full, non-contracted graph. The contracted graph can be generated by passing the result to dodgr_contract_graph.

Usage

dodgr_load_streetnet(filename)

Arguments

filename

Name (with optional full path) of file to be loaded.

Value

The loaded street network.

See also

Examples

net <- weight_streetnet (hampi)
f <- file.path (tempdir (), "streetnet.Rds")
dodgr_save_streetnet (net, f)
clear_dodgr_cache () # rm cached objects from tempdir
# at some later time, or in a new R session:
net <- dodgr_load_streetnet (f)