dodgr_streetnet

Use the osmdata package to extract the street network for a given location. For routing between a given set of points (passed as pts), the bbox argument may be omitted, in which case a bounding box will be constructed by expanding the range of pts by the relative amount of expand.

dodgr_streetnet(bbox, pts, expand = 0.05, quiet = TRUE)

Arguments

bbox
Bounding box as vector or matrix of coordinates, or location name. Passed to osmdata::getbb.

pts
List of points presumably containing spatial coordinates

expand
Relative factor by which street network should extend beyond limits defined by pts (only if bbox not given).

quiet
If FALSE, display progress messages

Value

A Simple Features (sf) object with coordinates of all lines in the street network.

Examples

# NOT RUN {
streetnet <- dodgr_streetnet ("hampi india", expand = 0)
# convert to form needed for \code{dodgr} functions:
graph <- weight_streetnet (streetnet)
nrow (graph) # 5,742 edges
# Alternative ways of extracting street networks by using a small selection of
# graph vertices to define bounding box:
verts <- dodgr_vertices (graph)
verts <- verts [sample (nrow (verts), size = 200), ]
streetnet <- dodgr_streetnet (pts = verts, expand = 0)
graph <- weight_streetnet (streetnet)
nrow (graph)
# This will generally have many more rows because most street networks include
# streets that extend considerably beyond the specified bounding box.
# }