Extract vertices of graph, including spatial coordinates if included.

Extract vertices of graph, including spatial coordinates if included.

Usage

dodgr_vertices(graph)

Arguments

graph

A flat table of graph edges. Must contain columns labelled from and to, or start and stop. May also contain similarly labelled columns of spatial coordinates (for example from_x) or stop_lon).

Value

A data.frame of vertices with unique numbers (n).

Note

Values of n are 0-indexed

See also

Examples

graph <- weight_streetnet (hampi)
v <- dodgr_vertices (graph)