Compare timings of different sort heaps for a given input graph.

Perform timing comparison between different kinds of heaps as well as with equivalent routines from the igraph package. To do this, a random sub-graph containing a defined number of vertices is first selected. Alternatively, this random sub-graph can be pre-generated with the dodgr_sample function and passed directly.

Usage

compare_heaps(graph, nverts = 100, replications = 2)

Arguments

graph

data.frame object representing the network graph (or a sub-sample selected with dodgr_sample)

nverts

Number of vertices used to generate random sub-graph. If a non-numeric value is given, the whole graph will be used.

replications

Number of replications to be used in comparison

Value

Result of bench::mark comparison.

See also

Examples

graph <- weight_streetnet (hampi)
if (FALSE) { # \dontrun{
compare_heaps (graph, nverts = 1000, replications = 1)
} # }