dodgr_isochrones

Isochrones from specified points. Function is fully vectorized to calculate accept vectors of central points and vectors of isodistance thresholds.

dodgr_isochrones(graph, from = NULL, tlim = NULL, heap = "BHeap")

Arguments

graph
data.frame or equivalent object representing the network graph (see Notes)

from
Vector or matrix of points from which isodistances are to be calculated (see Notes)

tlim
Desired limits of isochrones in seconds

heap
Type of heap to use in priority queue. Options include Fibonacci Heap (default; FHeap), Binary Heap (BHeap), Radix, Trinomial Heap (TriHeap), Extended Trinomial Heap (TriHeapExt, and 2-3 Heap (Heap23).

Value

A single data.frame of isochrones as points sorted anticlockwise around each origin (from) point, with columns denoting the from points and tlim value(s). The isochrones are given as id values of the series of points from each from point at the specified isodistance.

Examples

# NOT RUN {
# Use osmdata package to extract 'SC'-format data:
library (osmdata)
dat <- opq ("hampi india") %>%
    add_osm_feature (key = "highway") %>%
    osmdata_sc ()
graph <- weight_streetnet (dat)
from <- sample (graph$from_id, size = 100)
tlim <- c (5, 10, 20, 30, 60) * 60 # times in seconds
x <- dodgr_isochrones (graph, from = from, tlim)
# }