Force weight_streetnet to use geodesic distances.

Distances by default are Mapbox "cheap" distances if maximal network distances are < 100km, otherwise Haversine distances. Calling this function forces all calls to weight_streetnet from that point on to use geodesic distances. These are more computationally expensive to calculate, and weighting networks will likely take more time.

Usage

dodgr_streetnet_geodesic(unset = FALSE)

Arguments

unset

Calling this function with unset = TRUE reverts distance calculations to those described above, rather than geodesic.

Value

Nothing; the function is called for its side-effect only of setting distance calculations to geodesic.

See also

Examples

net0 <- weight_streetnet (hampi) # Default "cheap" method
dodgr_streetnet_geodesic ()
net1 <- weight_streetnet (hampi)
cor (net0$d, net1$d) # Strongly correlated, but not perfect
#> [1] 0.9999972
max (abs (net0$d - net1$d)) # in metres
#> [1] 1.622769