match_points_to_graph

Alias for match_pts_to_graph

Usage

match_points_to_graph(graph, xy, connected = FALSE)

Arguments

graph

A dodgr graph with spatial coordinates, such as a dodgr_streetnet object.

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

Value

A vector index matching the xy coordinates to nearest edges. For bi-directional edges, only one match is returned, and it is up to the user to identify and suitably process matching edge pairs.

See also

Examples

graph <- weight_streetnet (hampi, wt_profile = "foot")
# Then generate some random points to match to graph
verts <- dodgr_vertices (graph)
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
edges <- match_pts_to_graph (graph, xy)
graph [edges, ] # The edges of the graph closest to `xy`
#>        geom_num edge_id    from_id from_lon from_lat      to_id   to_lon
#> 6232        203    6232 1204772781 76.39982 15.35266 3526242955 76.39943
#> 3395         89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 3395.1       89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 58            1      58  313796397 76.48639 15.33916 2398957978 76.48626
#> 6256        203    6256 1204772830 76.42729 15.35071 1388482473 76.42580
#> 6256.1      203    6256 1204772830 76.42729 15.35071 1388482473 76.42580
#> 5926        183    5926  676635868 76.47446 15.30734 2398957516 76.47461
#> 6200        203    6200 7769271419 76.37263 15.34499 6025347259 76.37261
#> 3558         90    3558 2588155715 76.43778 15.33045 7794286114 76.43793
#> 3396         89    3396 2588146107 76.42349 15.31746 2588119056 76.42341
#>          to_lat          d d_weighted      highway    way_id component
#> 6232   15.35267  41.979318  83.958635      primary 835018468         2
#> 3395   15.31746  33.008446  41.260558 unclassified 252786290         1
#> 3395.1 15.31746  33.008446  41.260558 unclassified 252786290         1
#> 58     15.33917  13.655700  13.655700         path  28565950         1
#> 6256   15.35076 160.538138 321.076277      primary 835018468         2
#> 6256.1 15.35076 160.538138 321.076277      primary 835018468         2
#> 5926   15.30765  37.851136  75.702272      primary 652570479         1
#> 6200   15.34499   2.646247   5.292493      primary 835018468         2
#> 3558   15.33046  15.653674  16.477552        track 252787544         1
#> 3396   15.31717  33.008446  41.260558 unclassified 252786290         1
#>              time time_weighted
#> 6232    30.225109     60.450217
#> 3395    23.766081     29.707601
#> 3395.1  23.766081     29.707601
#> 58       9.832104      9.832104
#> 6256   115.587460    231.174919
#> 6256.1 115.587460    231.174919
#> 5926    27.252818     54.505636
#> 6200     1.905298      3.810595
#> 3558    11.270645     11.863837
#> 3396    23.766081     29.707601