match_points_to_graph

Alias for match_points_to_graph

match_points_to_graph(verts, xy, connected = FALSE)

Arguments

verts
A data.frame of vertices obtained from dodgr_vertices(graph).

xy
coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected
Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

Value

A vector index into verts

Examples

net <- weight_streetnet (hampi, wt_profile = "foot")
verts <- dodgr_vertices (net)
# Then generate some random points to match to graph
npts <- 10
xy <- data.frame (
                  x = min (verts$x) + runif (npts) * diff (range (verts$x)),
                  y = min (verts$y) + runif (npts) * diff (range (verts$y))
                  )
pts <- match_pts_to_graph (verts, xy)
pts # an index into verts

#>  [1]  468  760  398 1776  450  485  486  738  750  582

pts <- verts$id [pts]
pts # names of those vertices

#>  [1] "1691496070" "1204772780" "1128374373" "2588119056" "2398957939"
#>  [6] "1161117586" "1161117652" "6025347259" "1204772759" "3544918711"