Alias for match_pts_to_graph

Match spatial points to the edges of a spatial graph, through finding the edge with the closest perpendicular intersection. NOTE: Intersections are calculated geometrically, and presume planar geometry. It is up to users of projected geometrical data, such as those within a dodgr_streetnet object, to ensure that either: (i) Data span an sufficiently small area that errors from presuming planar geometry may be ignored; or (ii) Data are re-projected to an equivalent planar geometry prior to calling this routine.

Usage

match_points_to_graph(graph, xy, connected = FALSE)

Arguments

graph

A dodgr graph with spatial coordinates, such as a dodgr_streetnet object.

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

Value

For distances = FALSE (default), a vector index matching the xy coordinates to nearest edges. For bi-directional edges, only one match is returned, and it is up to the user to identify and suitably process matching edge pairs. For 'distances = TRUE', a 'data.frame' of four columns:

"index" The index of closest edges in "graph", as described above.

"d_signed" The perpendicular distance from ech point to the nearest edge, with negative distances denoting points to the left of edges, and positive distances denoting points to the right. Distances of zero denote points lying precisely on the line of an edge (potentially including cases where nearest point of bisection lies beyond the actual edge).

"x" The x-coordinate of the point of intersection.

"y" The y-coordinate of the point of intersection.

See also

Examples

graph <- weight_streetnet (hampi, wt_profile = "foot")
# Then generate some random points to match to graph
verts <- dodgr_vertices (graph)
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
edges <- match_pts_to_graph (graph, xy)
graph [edges, ] # The edges of the graph closest to `xy`
#>        geom_num edge_id    from_id from_lon from_lat      to_id   to_lon
#> 3395         89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 6251        203    6251 6597298435 76.42010 15.35136 6597300510 76.42343
#> 5917        183    5917 2419196948 76.47589 15.30930 2398957539 76.47550
#> 3395.1       89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 6253        203    6253 6597300510 76.42343 15.35101 1388482473 76.42580
#> 5915        183    5915  313796433 76.47604 15.30953 2419196948 76.47589
#> 6261        203    6261 3921525779 76.43165 15.35029 1204772909 76.43202
#> 4705        133    4705 2398957616 76.46775 15.31425 2398957612 76.46790
#> 5301        158    5301 3921524453 76.44483 15.34252 3921524452 76.44502
#> 6201        203    6201 7769271419 76.37263 15.34499 7769190961 76.38203
#>          to_lat          d d_weighted      highway    way_id component
#> 3395   15.31746   33.00845   41.26056 unclassified 252786290         1
#> 6251   15.35101  359.82586  719.65173      primary 835018468         2
#> 5917   15.30880   69.14107  138.28215      primary 652570479         1
#> 3395.1 15.31746   33.00845   41.26056 unclassified 252786290         1
#> 6253   15.35076  255.50004  511.00008      primary 835018468         2
#> 5915   15.30930   29.48529   58.97058      primary 652570479         1
#> 6261   15.35025   39.68515   79.37029      primary 835018468         2
#> 4705   15.31419   16.95158   28.25263    secondary 327102382         1
#> 5301   15.34246   20.81961   21.91537        track 388983289         2
#> 6201   15.34708 1035.68141 2071.36283      primary 835018468         2
#>             time time_weighted
#> 3395    23.76608      29.70760
#> 6251   259.07462     518.14924
#> 5917    49.78157      99.56314
#> 3395.1  23.76608      29.70760
#> 6253   183.96003     367.92006
#> 5915    21.22941      42.45882
#> 6261    28.57331      57.14661
#> 4705    12.20514      20.34189
#> 5301    14.99012      15.77907
#> 6201   745.69062    1491.38124