Alias for match_pts_to_graph

Match spatial points to the edges of a spatial graph, through finding the edge with the closest perpendicular intersection. NOTE: Intersections are calculated geometrically, and presume planar geometry. It is up to users of projected geometrical data, such as those within a dodgr_streetnet object, to ensure that either: (i) Data span an sufficiently small area that errors from presuming planar geometry may be ignored; or (ii) Data are re-projected to an equivalent planar geometry prior to calling this routine.

Usage

match_points_to_graph(graph, xy, connected = FALSE)

Arguments

graph

A dodgr graph with spatial coordinates, such as a dodgr_streetnet object.

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

Value

For 'distances = FALSE' (default), a vector index matching the xy

coordinates to nearest edges. For bi-directional edges, only one match is returned, and it is up to the user to identify and suitably process matching edge pairs. For 'distances = TRUE', a 'data.frame' of two columns:

"index" The index of closest edges in "graph", as described above.

"d_signed" The perpendicular distance from ech point to the nearest edge, with negative distances denoting points to the left of edges, and positive distances denoting points to the right. Distances of zero denote points lying precisely on the line of an edge (potentially including cases where nearest point of bisection lies beyond the actual edge).

See also

Examples

graph <- weight_streetnet (hampi, wt_profile = "foot")
# Then generate some random points to match to graph
verts <- dodgr_vertices (graph)
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
edges <- match_pts_to_graph (graph, xy)
graph [edges, ] # The edges of the graph closest to `xy`
#>      geom_num edge_id    from_id from_lon from_lat      to_id   to_lon   to_lat
#> 2995       82    2995 2195424984 76.45210 15.35248 2195424982 76.45240 15.35242
#> 5847      183    5847 7806119996 76.48527 15.32070  676875269 76.48495 15.32047
#> 2529       70    2529 7799711079 76.48111 15.33374 7799711078 76.48117 15.33374
#> 2671       70    2671 7799711113 76.48881 15.32998 4474520506 76.48880 15.32986
#> 6247      203    6247 1204772780 76.40951 15.35236 1204772868 76.41809 15.35157
#> 461        16     461  339574170 76.45255 15.33357  339574171 76.45226 15.33344
#> 5245      157    5245 5351564645 76.43569 15.34649 5351564647 76.43545 15.34650
#> 6253      203    6253 6597300510 76.42343 15.35101 1388482473 76.42580 15.35076
#> 3463       89    3463 7793361780 76.42942 15.31824 2588146149 76.43034 15.31818
#> 6237      203    6237 1388483320 76.40250 15.35238 1388482647 76.40387 15.35230
#>               d  d_weighted      highway    way_id component       time
#> 2995  32.734999   40.918749 unclassified 209318354         2  23.569200
#> 5847  42.932456   85.864912      primary 652570479         1  30.911368
#> 2529   6.283018    6.283018         path 123463598         1   4.523773
#> 2671  13.099850   13.099850         path 123463598         1   9.431892
#> 6247 925.873570 1851.747140      primary 835018468         2 666.628970
#> 461   34.760941   36.590464        track  30704676         1  25.027878
#> 5245  25.476002   26.816844        track 388983220         2  18.342721
#> 6253 255.500041  511.000082      primary 835018468         2 183.960030
#> 3463  99.273269  124.091586 unclassified 252786290         1  71.476754
#> 6237 146.939045  293.878090      primary 835018468         2 105.796112
#>      time_weighted
#> 2995     29.461499
#> 5847     61.822737
#> 2529      4.523773
#> 2671      9.431892
#> 6247   1333.257941
#> 461      26.345134
#> 5245     19.308128
#> 6253    367.920059
#> 3463     89.345942
#> 6237    211.592225