Alias for match_pts_to_graph

Match spatial points to the edges of a spatial graph, through finding the edge with the closest perpendicular intersection. NOTE: Intersections are calculated geometrically, and presume planar geometry. It is up to users of projected geometrical data, such as those within a dodgr_streetnet object, to ensure that either: (i) Data span an sufficiently small area that errors from presuming planar geometry may be ignored; or (ii) Data are re-projected to an equivalent planar geometry prior to calling this routine.

Usage

match_points_to_graph(graph, xy, connected = FALSE)

Arguments

graph

A dodgr graph with spatial coordinates, such as a dodgr_streetnet object.

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

Value

For distances = FALSE (default), a vector index matching the xy coordinates to nearest edges. For bi-directional edges, only one match is returned, and it is up to the user to identify and suitably process matching edge pairs. For 'distances = TRUE', a 'data.frame' of four columns:

"index" The index of closest edges in "graph", as described above.

"d_signed" The perpendicular distance from ech point to the nearest edge, with negative distances denoting points to the left of edges, and positive distances denoting points to the right. Distances of zero denote points lying precisely on the line of an edge (potentially including cases where nearest point of bisection lies beyond the actual edge).

"x" The x-coordinate of the point of intersection.

"y" The y-coordinate of the point of intersection.

See also

Examples

graph <- weight_streetnet (hampi, wt_profile = "foot")
# Then generate some random points to match to graph
verts <- dodgr_vertices (graph)
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
edges <- match_pts_to_graph (graph, xy)
graph [edges, ] # The edges of the graph closest to `xy`
#>        geom_num edge_id    from_id from_lon from_lat      to_id   to_lon
#> 3395         89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 3863         90    3863 2588155705 76.45578 15.32697 2398957796 76.45648
#> 3395.1       89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 6201        203    6201 7769271419 76.37263 15.34499 7769190961 76.38203
#> 3083         82    3083 2195424961 76.46392 15.35156 2195424963 76.46400
#> 6207        203    6207 1388481924 76.38383 15.34743 1388482509 76.38860
#> 5831        183    5831 1128374373 76.48920 15.32237 6161027522 76.48787
#> 3395.2       89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 2837         72    2837 1398747940 76.44880 15.35303 1398748037 76.44852
#> 3507         89    3507 5351820917 76.44137 15.32000 5351820918 76.44189
#>          to_lat          d d_weighted      highway    way_id component
#> 3395   15.31746   33.00845   41.26056 unclassified 252786290         1
#> 3863   15.32746   92.78039   97.66357        track 252787544         1
#> 3395.1 15.31746   33.00845   41.26056 unclassified 252786290         1
#> 6201   15.34708 1035.68141 2071.36283      primary 835018468         2
#> 3083   15.35168   15.60411   19.50514 unclassified 209318354         2
#> 6207   15.34745  511.92811 1023.85621      primary 835018468         2
#> 5831   15.32197  149.92249  299.84498      primary 652570479         1
#> 3395.2 15.31746   33.00845   41.26056 unclassified 252786290         1
#> 2837   15.35307   30.36742   37.95927 unclassified 126094049         2
#> 3507   15.32011   57.03688   71.29610 unclassified 252786290         1
#>             time time_weighted
#> 3395    23.76608      29.70760
#> 3863    66.80188      70.31777
#> 3395.1  23.76608      29.70760
#> 6201   745.69062    1491.38124
#> 3083    11.23496      14.04370
#> 6207   368.58824     737.17647
#> 5831   107.94419     215.88839
#> 3395.2  23.76608      29.70760
#> 2837    21.86454      27.33068
#> 3507    41.06655      51.33319