Alias for match_pts_to_graph

Match spatial points to the edges of a spatial graph, through finding the edge with the closest perpendicular intersection. NOTE: Intersections are calculated geometrically, and presume planar geometry. It is up to users of projected geometrical data, such as those within a dodgr_streetnet object, to ensure that either: (i) Data span an sufficiently small area that errors from presuming planar geometry may be ignored; or (ii) Data are re-projected to an equivalent planar geometry prior to calling this routine.

Usage

match_points_to_graph(graph, xy, connected = FALSE)

Arguments

graph

A dodgr graph with spatial coordinates, such as a dodgr_streetnet object.

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

Value

For 'distances = FALSE' (default), a vector index matching the xy

coordinates to nearest edges. For bi-directional edges, only one match is returned, and it is up to the user to identify and suitably process matching edge pairs. For 'distances = TRUE', a 'data.frame' of two columns:

"index" The index of closest edges in "graph", as described above.

"d_signed" The perpendicular distance from ech point to the nearest edge, with negative distances denoting points to the left of edges, and positive distances denoting points to the right. Distances of zero denote points lying precisely on the line of an edge (potentially including cases where nearest point of bisection lies beyond the actual edge).

See also

Examples

graph <- weight_streetnet (hampi, wt_profile = "foot")
# Then generate some random points to match to graph
verts <- dodgr_vertices (graph)
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
edges <- match_pts_to_graph (graph, xy)
graph [edges, ] # The edges of the graph closest to `xy`
#>      geom_num edge_id    from_id from_lon from_lat      to_id   to_lon   to_lat
#> 6253      203    6253 6597300510 76.42343 15.35101 1388482473 76.42580 15.35076
#> 3463       89    3463 7793361780 76.42942 15.31824 2588146149 76.43034 15.31818
#> 6237      203    6237 1388483320 76.40250 15.35238 1388482647 76.40387 15.35230
#> 21          1      21 2398958016 76.48068 15.34096  313796394 76.48195 15.34061
#> 3395       89    3395 2588119056 76.42341 15.31717 2588146107 76.42349 15.31746
#> 5299      158    5299 3921525057 76.44464 15.34258 3921524453 76.44483 15.34252
#> 6259      203    6259 1388481903 76.43011 15.35046 3921525779 76.43165 15.35029
#> 57          1      57 2398957978 76.48626 15.33917  313796397 76.48639 15.33916
#> 5675      171    5675 5974426256 76.44354 15.31398 5351820892 76.44350 15.31399
#> 259         4     259 1676740773 76.46446 15.33427 4474520824 76.46335 15.33453
#>               d d_weighted      highway    way_id component       time
#> 6253 255.500041 511.000082      primary 835018468         2 183.960030
#> 3463  99.273269 124.091586 unclassified 252786290         1  71.476754
#> 6237 146.939045 293.878090      primary 835018468         2 105.796112
#> 21   142.394986 142.394986         path  28565950         1 102.524390
#> 3395  33.008446  41.260558 unclassified 252786290         1  23.766081
#> 5299  21.302084  22.423247        track 388983289         2  15.337501
#> 6259 166.130322 332.260645      primary 835018468         2 119.613832
#> 57    13.655700  13.655700         path  28565950         1   9.832104
#> 5675   4.550499   5.688124 unclassified 554572321         1   3.276359
#> 259  123.271145 136.967938      service  30659140         1  88.755224
#>      time_weighted
#> 6253    367.920059
#> 3463     89.345942
#> 6237    211.592225
#> 21      102.524390
#> 3395     29.707601
#> 5299     16.144738
#> 6259    239.227664
#> 57        9.832104
#> 5675      4.095449
#> 259      98.616916