Alias for match_pts_to_graph

Match spatial points to the edges of a spatial graph, through finding the edge with the closest perpendicular intersection. NOTE: Intersections are calculated geometrically, and presume planar geometry. It is up to users of projected geometrical data, such as those within a dodgr_streetnet object, to ensure that either: (i) Data span an sufficiently small area that errors from presuming planar geometry may be ignored; or (ii) Data are re-projected to an equivalent planar geometry prior to calling this routine.

Usage

match_points_to_graph(graph, xy, connected = FALSE)

Arguments

graph

A dodgr graph with spatial coordinates, such as a dodgr_streetnet object.

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

Value

For 'distances = FALSE' (default), a vector index matching the xy

coordinates to nearest edges. For bi-directional edges, only one match is returned, and it is up to the user to identify and suitably process matching edge pairs. For 'distances = TRUE', a 'data.frame' of two columns:

"index" The index of closest edges in "graph", as described above.

"d_signed" The perpendicular distance from ech point to the nearest edge, with negative distances denoting points to the left of edges, and positive distances denoting points to the right. Distances of zero denote points lying precisely on the line of an edge (potentially including cases where nearest point of bisection lies beyond the actual edge).

See also

Examples

graph <- weight_streetnet (hampi, wt_profile = "foot")
# Then generate some random points to match to graph
verts <- dodgr_vertices (graph)
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
edges <- match_pts_to_graph (graph, xy)
graph [edges, ] # The edges of the graph closest to `xy`
#>        geom_num edge_id    from_id from_lon from_lat      to_id   to_lon
#> 6233        203    6233 1204772781 76.39982 15.35266 1204772873 76.40114
#> 3395         89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 3395.1       89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 57            1      57 2398957978 76.48626 15.33917  313796397 76.48639
#> 6257        203    6257 1204772830 76.42729 15.35071 1388481903 76.43011
#> 6257.1      203    6257 1204772830 76.42729 15.35071 1388481903 76.43011
#> 5925        183    5925 2398957516 76.47461 15.30765  676635868 76.47446
#> 6201        203    6201 7769271419 76.37263 15.34499 7769190961 76.38203
#> 3557         90    3557 7794286114 76.43793 15.33046 2588155715 76.43778
#> 3395.2       89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#>          to_lat          d d_weighted      highway    way_id component
#> 6233   15.35250  142.94319  285.88639      primary 835018468         2
#> 3395   15.31746   33.00845   41.26056 unclassified 252786290         1
#> 3395.1 15.31746   33.00845   41.26056 unclassified 252786290         1
#> 57     15.33916   13.65570   13.65570         path  28565950         1
#> 6257   15.35046  304.13301  608.26602      primary 835018468         2
#> 6257.1 15.35046  304.13301  608.26602      primary 835018468         2
#> 5925   15.30734   37.85114   75.70227      primary 652570479         1
#> 6201   15.34708 1035.68141 2071.36283      primary 835018468         2
#> 3557   15.33045   15.65367   16.47755        track 252787544         1
#> 3395.2 15.31746   33.00845   41.26056 unclassified 252786290         1
#>              time time_weighted
#> 6233   102.919100    205.838200
#> 3395    23.766081     29.707601
#> 3395.1  23.766081     29.707601
#> 57       9.832104      9.832104
#> 6257   218.975766    437.951532
#> 6257.1 218.975766    437.951532
#> 5925    27.252818     54.505636
#> 6201   745.690618   1491.381236
#> 3557    11.270645     11.863837
#> 3395.2  23.766081     29.707601