Alias for match_pts_to_graph

Match spatial points to the edges of a spatial graph, through finding the edge with the closest perpendicular intersection. NOTE: Intersections are calculated geometrically, and presume planar geometry. It is up to users of projected geometrical data, such as those within a dodgr_streetnet object, to ensure that either: (i) Data span an sufficiently small area that errors from presuming planar geometry may be ignored; or (ii) Data are re-projected to an equivalent planar geometry prior to calling this routine.

Usage

match_points_to_graph(graph, xy, connected = FALSE)

Arguments

graph

A dodgr graph with spatial coordinates, such as a dodgr_streetnet object.

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

Value

For distances = FALSE (default), a vector index matching the xy

coordinates to nearest edges. For bi-directional edges, only one match is returned, and it is up to the user to identify and suitably process matching edge pairs. For 'distances = TRUE', a 'data.frame' of four columns:

"index" The index of closest edges in "graph", as described above.

"d_signed" The perpendicular distance from ech point to the nearest edge, with negative distances denoting points to the left of edges, and positive distances denoting points to the right. Distances of zero denote points lying precisely on the line of an edge (potentially including cases where nearest point of bisection lies beyond the actual edge).

"x" The x-coordinate of the point of intersection.

"y" The y-coordinate of the point of intersection.

See also

Examples

graph <- weight_streetnet (hampi, wt_profile = "foot")
# Then generate some random points to match to graph
verts <- dodgr_vertices (graph)
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
edges <- match_pts_to_graph (graph, xy)
graph [edges, ] # The edges of the graph closest to `xy`
#>        geom_num edge_id    from_id from_lon from_lat      to_id   to_lon
#> 6099        192    6099  339932428 76.47006 15.33396 2627465884 76.47009
#> 6239        203    6239 1388482647 76.40387 15.35230 1204772675 76.40457
#> 3395         89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 4267        110    4267 2631613055 76.46191 15.34002 2631613057 76.46204
#> 4717        133    4717  338905105 76.47080 15.31338 2398957595 76.47096
#> 4117         98    4117 1143452160 76.45984 15.34616 1143452081 76.45924
#> 5847        183    5847 7806119996 76.48527 15.32070  676875269 76.48495
#> 3395.1       89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 4811        136    4811 3456932090 76.47442 15.35063 6071931424 76.47425
#> 2525         70    2525 1376768789 76.48082 15.33366 1376769175 76.48104
#>          to_lat         d d_weighted      highway    way_id component     time
#> 6099   15.33288 119.03618  119.03618         path 684236862         1 85.70605
#> 6239   15.35226  75.74146  151.48292      primary 835018468         2 54.53385
#> 3395   15.31746  33.00845   41.26056 unclassified 252786290         1 23.76608
#> 4267   15.34002  14.12029   14.12029         path 257679733         2 10.16661
#> 4717   15.31333  18.16062   30.26770    secondary 327102382         1 13.07565
#> 4117   15.34599  66.43509   69.93168        track 257141719         2 47.83327
#> 5847   15.32047  42.93246   85.86491      primary 652570479         1 30.91137
#> 3395.1 15.31746  33.00845   41.26056 unclassified 252786290         1 23.76608
#> 4811   15.35063  17.92081   17.92081         path 338585806         2 12.90298
#> 2525   15.33373  25.03653   25.03653         path 123463598         1 18.02630
#>        time_weighted
#> 6099        85.70605
#> 6239       109.06770
#> 3395        29.70760
#> 4267        10.16661
#> 4717        21.79275
#> 4117        50.35081
#> 5847        61.82274
#> 3395.1      29.70760
#> 4811        12.90298
#> 2525        18.02630