match_pts_to_graph

Match spatial points to the edges of a spatial graph, through finding the edge with the closest perpendicular intersection. NOTE: Intersections are calculated geometrically, and presume planar geometry. It is up to users of projected geometrical data, such as those within a dodgr_streetnet object, to ensure that either: (i) Data span an sufficiently small area that errors from presuming planar geometry may be ignored; or (ii) Data are re-projected to an equivalent planar geometry prior to calling this routine.

Usage

match_pts_to_graph(graph, xy, connected = FALSE)

Arguments

graph

A dodgr graph with spatial coordinates, such as a dodgr_streetnet object.

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

Value

A vector index matching the xy coordinates to nearest edges. For bi-directional edges, only one match is returned, and it is up to the user to identify and suitably process matching edge pairs.

See also

Examples

graph <- weight_streetnet (hampi, wt_profile = "foot")
# Then generate some random points to match to graph
verts <- dodgr_vertices (graph)
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
edges <- match_pts_to_graph (graph, xy)
graph [edges, ] # The edges of the graph closest to `xy`
#>      geom_num edge_id    from_id from_lon from_lat      to_id   to_lon   to_lat
#> 6236      203    6236 1388483320 76.40250 15.35238 1204772873 76.40114 15.35250
#> 6226      203    6226 1204772759 76.39828 15.35260 1204772661 76.39213 15.35169
#> 3472       89    3472 7793361778 76.43309 15.31772 2588146136 76.43231 15.31795
#> 6248      203    6248 1204772868 76.41809 15.35157 1204772780 76.40951 15.35236
#> 6200      203    6200 7769271419 76.37263 15.34499 6025347259 76.37261 15.34499
#> 184         2     184 2398957683 76.47765 15.31646 2398957681 76.47763 15.31640
#> 3586       90    3586 2588155744 76.43550 15.33031 2588155764 76.43554 15.33031
#> 6199      203    6199 6025347259 76.37261 15.34499 7769271419 76.37263 15.34499
#> 6252      203    6252 6597300510 76.42343 15.35101 6597298435 76.42010 15.35136
#> 6202      203    6202 7769190961 76.38203 15.34708 7769271419 76.37263 15.34499
#>                d  d_weighted      highway    way_id component       time
#> 6236  146.561375  293.122749      primary 835018468         2 105.524190
#> 6226  667.359739 1334.719477      primary 835018468         2 480.499012
#> 3472   87.166514  108.958143 unclassified 252786290         1  62.759890
#> 6248  925.873570 1851.747140      primary 835018468         2 666.628970
#> 6200    2.646247    5.292493      primary 835018468         2   1.905298
#> 184     6.802001    6.802001         path  30643853         1  12.243601
#> 3586    4.370737    4.600776        track 252787544         1   3.146931
#> 6199    2.646247    5.292493      primary 835018468         2   1.905298
#> 6252  359.825864  719.651728      primary 835018468         2 259.074622
#> 6202 1035.681414 2071.362827      primary 835018468         2 745.690618
#>      time_weighted
#> 6236    211.048380
#> 6226    960.998024
#> 3472     78.449863
#> 6248   1333.257941
#> 6200      3.810595
#> 184      12.243601
#> 3586      3.312559
#> 6199      3.810595
#> 6252    518.149244
#> 6202   1491.381236