match_pts_to_graph

Match spatial points to a spatial graph which contains vertex coordinates

match_pts_to_graph(verts, xy, connected = FALSE)

Arguments

verts

A data.frame of vertices obtained from dodgr_vertices(graph).

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

Value

A vector index into verts

See also

Examples

net <- weight_streetnet (hampi, wt_profile = "foot")
verts <- dodgr_vertices (net)
# Then generate some random points to match to graph
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
pts <- match_pts_to_graph (verts, xy)
pts # an index into verts
#>  [1] 1709 2590 3053 1704 1908  652 2922 3055 3069  101
pts <- verts$id [pts]
pts # names of those vertices
#>  [1] "2588146138" "5351564644" "7769190961" "2588119056" "7794251336"
#>  [6] "571423205"  "2419196948" "1388481924" "1204772873" "2398957695"