match_pts_to_graph

Match spatial points to a spatial graph which contains vertex coordindates

match_pts_to_graph(verts, xy, connected = FALSE)

Arguments

verts
A data.frame of vertices obtained from dodgr_vertices(graph).

xy
coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected
Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routine routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

Value

A vector index into verts

Examples

net <- weight_streetnet (hampi, wt_profile = "foot")

#> The following highway types are present in data yet lack corresponding weight_profile values: living_street, 

verts <- dodgr_vertices (net)
# Then generate some random points to match to graph
npts <- 10
xy <- data.frame (
                  x = min (verts$x) + runif (npts) * diff (range (verts$x)),
                  y = min (verts$y) + runif (npts) * diff (range (verts$y))
                  )
pts <- match_pts_to_graph (verts, xy)
pts # an index into verts

#>  [1]  540 1782  768  516  768  760  484  782  727  726

pts <- verts$id [pts]
pts # names of those vertices

#>  [1] "677638828"  "2588119056" "1204772855" "1691495727" "1204772855"
#>  [6] "1204772804" "1161117652" "1204772868" "1412452580" "1388482269"