Match spatial points to the edges of a spatial graph.

Match spatial points to the edges of a spatial graph, through finding the edge with the closest perpendicular intersection. NOTE: Intersections are calculated geometrically, and presume planar geometry. It is up to users of projected geometrical data, such as those within a dodgr_streetnet object, to ensure that either: (i) Data span an sufficiently small area that errors from presuming planar geometry may be ignored; or (ii) Data are re-projected to an equivalent planar geometry prior to calling this routine.

Usage

match_pts_to_graph(graph, xy, connected = FALSE, distances = FALSE)

Arguments

graph

A dodgr graph with spatial coordinates, such as a dodgr_streetnet object.

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

distances

If TRUE, return a 'data.frame' object with 'index' column as described in return value; and additional columns with perpendicular distance to nearest edge in graph, and coordinates of points of intersection. See description of return value for details.

Value

For distances = FALSE (default), a vector index matching the xy coordinates to nearest edges. For bi-directional edges, only one match is returned, and it is up to the user to identify and suitably process matching edge pairs. For 'distances = TRUE', a 'data.frame' of four columns:

"index" The index of closest edges in "graph", as described above.

"d_signed" The perpendicular distance from ech point to the nearest edge, with negative distances denoting points to the left of edges, and positive distances denoting points to the right. Distances of zero denote points lying precisely on the line of an edge (potentially including cases where nearest point of bisection lies beyond the actual edge).

"x" The x-coordinate of the point of intersection.

"y" The y-coordinate of the point of intersection.

See also

Examples

graph <- weight_streetnet (hampi, wt_profile = "foot")
# Then generate some random points to match to graph
verts <- dodgr_vertices (graph)
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
edges <- match_pts_to_graph (graph, xy)
graph [edges, ] # The edges of the graph closest to `xy`
#>        geom_num edge_id    from_id from_lon from_lat      to_id   to_lon
#> 3405         89    3405 2588146138 76.42439 15.31886 2588146016 76.42477
#> 6225        203    6225 1204772661 76.39213 15.35169 1204772759 76.39828
#> 6221        203    6221 1204772855 76.39160 15.35146 1388481881 76.39199
#> 6221.1      203    6221 1204772855 76.39160 15.35146 1388481881 76.39199
#> 57            1      57 2398957978 76.48626 15.33917  313796397 76.48639
#> 955          34     955 2398957751 76.45195 15.32208 2398957754 76.45238
#> 5245        157    5245 5351564645 76.43569 15.34649 5351564647 76.43545
#> 3279         85    3279 2195425073 76.45951 15.35706 2195425070 76.46011
#> 6199        203    6199 6025347259 76.37261 15.34499 7769271419 76.37263
#> 2881         72    2881 1398748033 76.44181 15.35615 1398748009 76.44158
#>          to_lat          d  d_weighted      highway    way_id component
#> 3405   15.31911  48.825372   61.031715 unclassified 252786290         1
#> 6225   15.35260 666.247353 1332.494706      primary 835018468         2
#> 6221   15.35165  46.168705   92.337410      primary 835018468         2
#> 6221.1 15.35165  46.168705   92.337410      primary 835018468         2
#> 57     15.33916  13.630942   13.630942         path  28565950         1
#> 955    15.32231  53.935636   89.892726    secondary  53626074         1
#> 5245   15.34650  25.429620   26.768021        track 388983220         2
#> 3279   15.35689  66.868015   66.868015         path 209318367         2
#> 6199   15.34499   2.642092    5.284183      primary 835018468         2
#> 2881   15.35644  39.547467   49.434334 unclassified 126094049         2
#>              time time_weighted
#> 3405    35.154268     43.942835
#> 6225   479.698094    959.396188
#> 6221    33.241468     66.482935
#> 6221.1  33.241468     66.482935
#> 57       9.814278      9.814278
#> 955     38.833658     64.722763
#> 5245    18.309326     19.272975
#> 3279    48.144971     48.144971
#> 6199     1.902306      3.804612
#> 2881    28.474176     35.592721