Match spatial points to the edges of a spatial graph.

Match spatial points to the edges of a spatial graph, through finding the edge with the closest perpendicular intersection. NOTE: Intersections are calculated geometrically, and presume planar geometry. It is up to users of projected geometrical data, such as those within a dodgr_streetnet object, to ensure that either: (i) Data span an sufficiently small area that errors from presuming planar geometry may be ignored; or (ii) Data are re-projected to an equivalent planar geometry prior to calling this routine.

Usage

match_pts_to_graph(graph, xy, connected = FALSE, distances = FALSE)

Arguments

graph

A dodgr graph with spatial coordinates, such as a dodgr_streetnet object.

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

distances

If TRUE, return a 'data.frame' object with 'index' column as described in return value; and additional columns with perpendicular distance to nearest edge in graph, and coordinates of points of intersection. See description of return value for details.

Value

For distances = FALSE (default), a vector index matching the xy

coordinates to nearest edges. For bi-directional edges, only one match is returned, and it is up to the user to identify and suitably process matching edge pairs. For 'distances = TRUE', a 'data.frame' of four columns:

"index" The index of closest edges in "graph", as described above.

"d_signed" The perpendicular distance from ech point to the nearest edge, with negative distances denoting points to the left of edges, and positive distances denoting points to the right. Distances of zero denote points lying precisely on the line of an edge (potentially including cases where nearest point of bisection lies beyond the actual edge).

"x" The x-coordinate of the point of intersection.

"y" The y-coordinate of the point of intersection.

See also

Examples

graph <- weight_streetnet (hampi, wt_profile = "foot")
# Then generate some random points to match to graph
verts <- dodgr_vertices (graph)
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
edges <- match_pts_to_graph (graph, xy)
graph [edges, ] # The edges of the graph closest to `xy`
#>        geom_num edge_id    from_id from_lon from_lat      to_id   to_lon
#> 5517        169    5517 4506481406 76.46174 15.31509 4506481407 76.46154
#> 6239        203    6239 1388482647 76.40387 15.35230 1204772675 76.40457
#> 6247        203    6247 1204772780 76.40951 15.35236 1204772868 76.41809
#> 6241        203    6241 1204772675 76.40457 15.35226 1204772877 76.40606
#> 3253         85    3253 2195425109 76.45403 15.35955 2195425106 76.45432
#> 6249        203    6249 1204772868 76.41809 15.35157 6597298435 76.42010
#> 3473         89    3473 7793361778 76.43309 15.31772 7793361779 76.43379
#> 5851        183    5851  245631169 76.48423 15.32007 7806119997 76.48362
#> 6241.1      203    6241 1204772675 76.40457 15.35226 1204772877 76.40606
#> 1311         35    1311 1206252905 76.49212 15.35451  571423209 76.49232
#>          to_lat         d d_weighted      highway    way_id component      time
#> 5517   15.31516  23.36460   24.59432        track 454003193         1  16.82252
#> 6239   15.35226  75.74146  151.48292      primary 835018468         2  54.53385
#> 6247   15.35157 925.87357 1851.74714      primary 835018468         2 666.62897
#> 6241   15.35237 160.21420  320.42840      primary 835018468         2 115.35422
#> 3253   15.35937  37.11878   37.11878         path 209318367         2  26.72552
#> 6249   15.35136 216.89113  433.78226      primary 835018468         2 156.16161
#> 3473   15.31763  75.91522   94.89403 unclassified 252786290         1  54.65896
#> 5851   15.31985  70.40027  140.80055      primary 652570479         1  50.68820
#> 6241.1 15.35237 160.21420  320.42840      primary 835018468         2 115.35422
#> 1311   15.35470  29.77597   59.55195      primary  53658844         2  21.43870
#>        time_weighted
#> 5517        17.70791
#> 6239       109.06770
#> 6247      1333.25794
#> 6241       230.70845
#> 3253        26.72552
#> 6249       312.32323
#> 3473        68.32370
#> 5851       101.37639
#> 6241.1     230.70845
#> 1311        42.87740