Match spatial points to the edges of a spatial graph.

Match spatial points to the edges of a spatial graph, through finding the edge with the closest perpendicular intersection. NOTE: Intersections are calculated geometrically, and presume planar geometry. It is up to users of projected geometrical data, such as those within a dodgr_streetnet object, to ensure that either: (i) Data span an sufficiently small area that errors from presuming planar geometry may be ignored; or (ii) Data are re-projected to an equivalent planar geometry prior to calling this routine.

Usage

match_pts_to_graph(graph, xy, connected = FALSE, distances = FALSE)

Arguments

graph

A dodgr graph with spatial coordinates, such as a dodgr_streetnet object.

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

distances

If TRUE, return a 'data.frame' object with 'index' column as described in return value; and additional columns with perpendicular distance to nearest edge in graph, and coordinates of points of intersection. See description of return value for details.

Value

For distances = FALSE (default), a vector index matching the xy coordinates to nearest edges. For bi-directional edges, only one match is returned, and it is up to the user to identify and suitably process matching edge pairs. For 'distances = TRUE', a 'data.frame' of four columns:

"index" The index of closest edges in "graph", as described above.

"d_signed" The perpendicular distance from ech point to the nearest edge, with negative distances denoting points to the left of edges, and positive distances denoting points to the right. Distances of zero denote points lying precisely on the line of an edge (potentially including cases where nearest point of bisection lies beyond the actual edge).

"x" The x-coordinate of the point of intersection.

"y" The y-coordinate of the point of intersection.

See also

Examples

graph <- weight_streetnet (hampi, wt_profile = "foot")
# Then generate some random points to match to graph
verts <- dodgr_vertices (graph)
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
edges <- match_pts_to_graph (graph, xy)
graph [edges, ] # The edges of the graph closest to `xy`
#>        geom_num edge_id    from_id from_lon from_lat      to_id   to_lon
#> 6209        203    6209 1388482509 76.38860 15.34745 1204772662 76.38895
#> 6211        203    6211 1204772662 76.38895 15.34748 1204772772 76.38938
#> 6203        203    6203 7769190961 76.38203 15.34708 1204772804 76.38329
#> 6201        203    6201 7769271419 76.37263 15.34499 7769190961 76.38203
#> 1149         35    1149 1388483505 76.45515 15.34924  977490091 76.45597
#> 6243        203    6243 1204772877 76.40606 15.35237 1204772708 76.40809
#> 4573        129    4573 2398957784 76.45984 15.32591 8615501939 76.45987
#> 6209.1      203    6209 1388482509 76.38860 15.34745 1204772662 76.38895
#> 3395         89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 3405         89    3405 2588146138 76.42439 15.31886 2588146016 76.42477
#>          to_lat          d d_weighted      highway    way_id component
#> 6209   15.34748   38.16886   76.33773      primary 835018468         2
#> 6211   15.34775   55.47203  110.94405      primary 835018468         2
#> 6203   15.34736  138.66664  277.33329      primary 835018468         2
#> 6201   15.34708 1035.68141 2071.36283      primary 835018468         2
#> 1149   15.34995  117.72114  235.44229      primary  53658844         2
#> 6243   15.35245  218.61497  437.22995      primary 835018468         2
#> 4573   15.32603   13.80307   23.00511    secondary 327102372         1
#> 6209.1 15.34748   38.16886   76.33773      primary 835018468         2
#> 3395   15.31746   33.00845   41.26056 unclassified 252786290         1
#> 3405   15.31911   48.82435   61.03044 unclassified 252786290         1
#>              time time_weighted
#> 6209    27.481581      54.96316
#> 6211    39.939860      79.87972
#> 6203    99.839984     199.67997
#> 6201   745.690618    1491.38124
#> 1149    84.759223     169.51845
#> 6243   157.402782     314.80556
#> 4573     9.938207      16.56368
#> 6209.1  27.481581      54.96316
#> 3395    23.766081      29.70760
#> 3405    35.153534      43.94192