Match spatial points to the edges of a spatial graph.

Match spatial points to the edges of a spatial graph, through finding the edge with the closest perpendicular intersection. NOTE: Intersections are calculated geometrically, and presume planar geometry. It is up to users of projected geometrical data, such as those within a dodgr_streetnet object, to ensure that either: (i) Data span an sufficiently small area that errors from presuming planar geometry may be ignored; or (ii) Data are re-projected to an equivalent planar geometry prior to calling this routine.

Usage

match_pts_to_graph(graph, xy, connected = FALSE, distances = FALSE)

Arguments

graph

A dodgr graph with spatial coordinates, such as a dodgr_streetnet object.

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

distances

If TRUE, return a 'data.frame' object with 'index' column as described in return value; and additional columns with perpendicular distance to nearest edge in graph, and coordinates of points of intersection. See description of return value for details.

Value

For distances = FALSE (default), a vector index matching the xy coordinates to nearest edges. For bi-directional edges, only one match is returned, and it is up to the user to identify and suitably process matching edge pairs. For 'distances = TRUE', a 'data.frame' of four columns:

"index" The index of closest edges in "graph", as described above.

"d_signed" The perpendicular distance from ech point to the nearest edge, with negative distances denoting points to the left of edges, and positive distances denoting points to the right. Distances of zero denote points lying precisely on the line of an edge (potentially including cases where nearest point of bisection lies beyond the actual edge).

"x" The x-coordinate of the point of intersection.

"y" The y-coordinate of the point of intersection.

See also

Examples

graph <- weight_streetnet (hampi, wt_profile = "foot")
# Then generate some random points to match to graph
verts <- dodgr_vertices (graph)
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
edges <- match_pts_to_graph (graph, xy)
graph [edges, ] # The edges of the graph closest to `xy`
#>        geom_num edge_id    from_id from_lon from_lat      to_id   to_lon
#> 1121         35    1121  977472830 76.45051 15.34541  977473072 76.45065
#> 3395         89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 6247        203    6247 1204772780 76.40951 15.35236 1204772868 76.41809
#> 4713        133    4713  338905101 76.46858 15.31411  338905103 76.46961
#> 6263        203    6263 1204772909 76.43202 15.35025 1204772735 76.43262
#> 6245        203    6245 1204772708 76.40809 15.35245 1204772780 76.40951
#> 315          10     315  980400557 76.45974 15.33539  339060019 76.45972
#> 4685        133    4685 1427115983 76.46557 15.31505 2398957629 76.46567
#> 2101         69    2101 1376768817 76.45828 15.31949 6414886574 76.45834
#> 3395.1       89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#>          to_lat          d d_weighted      highway    way_id component
#> 1121   15.34561  26.137835   52.27567      primary  53658844         2
#> 3395   15.31746  33.008446   41.26056 unclassified 252786290         1
#> 6247   15.35157 925.873570 1851.74714      primary 835018468         2
#> 4713   15.31386 114.530117  190.88353    secondary 327102382         1
#> 6263   15.35011  66.761819  133.52364      primary 835018468         2
#> 6245   15.35236 151.916453  303.83291      primary 835018468         2
#> 315    15.33555  17.811729   17.81173         path  30663412         1
#> 4685   15.31501  10.860469   18.10078    secondary 327102382         1
#> 2101   15.31955   8.910211   11.13776 unclassified 123463595         1
#> 3395.1 15.31746  33.008446   41.26056 unclassified 252786290         1
#>              time time_weighted
#> 1121    18.819241      37.63848
#> 3395    23.766081      29.70760
#> 6247   666.628970    1333.25794
#> 4713    82.461684     137.43614
#> 6263    48.068510      96.13702
#> 6245   109.379846     218.75969
#> 315     12.824445      12.82444
#> 4685     7.819538      13.03256
#> 2101     6.415352       8.01919
#> 3395.1  23.766081      29.70760