Match spatial points to the edges of a spatial graph.

Match spatial points to the edges of a spatial graph, through finding the edge with the closest perpendicular intersection. NOTE: Intersections are calculated geometrically, and presume planar geometry. It is up to users of projected geometrical data, such as those within a dodgr_streetnet object, to ensure that either: (i) Data span an sufficiently small area that errors from presuming planar geometry may be ignored; or (ii) Data are re-projected to an equivalent planar geometry prior to calling this routine.

Usage

match_pts_to_graph(graph, xy, connected = FALSE, distances = FALSE)

Arguments

graph

A dodgr graph with spatial coordinates, such as a dodgr_streetnet object.

xy

coordinates of points to be matched to the vertices, either as matrix or sf-formatted data.frame.

connected

Should points be matched to the same (largest) connected component of graph? If FALSE and these points are to be used for a dodgr routing routine (dodgr_dists, dodgr_paths, or dodgr_flows_aggregate), then results may not be returned if points are not part of the same connected component. On the other hand, forcing them to be part of the same connected component may decrease the spatial accuracy of matching.

distances

If TRUE, return a 'data.frame' object with 'index' column as described in return value; and additional 'dist' column with perpendicular distance to nearest edge in graph. See description of return value for details.

Value

For 'distances = FALSE' (default), a vector index matching the xy

coordinates to nearest edges. For bi-directional edges, only one match is returned, and it is up to the user to identify and suitably process matching edge pairs. For 'distances = TRUE', a 'data.frame' of two columns:

"index" The index of closest edges in "graph", as described above.

"d_signed" The perpendicular distance from ech point to the nearest edge, with negative distances denoting points to the left of edges, and positive distances denoting points to the right. Distances of zero denote points lying precisely on the line of an edge (potentially including cases where nearest point of bisection lies beyond the actual edge).

See also

Examples

graph <- weight_streetnet (hampi, wt_profile = "foot")
# Then generate some random points to match to graph
verts <- dodgr_vertices (graph)
npts <- 10
xy <- data.frame (
    x = min (verts$x) + runif (npts) * diff (range (verts$x)),
    y = min (verts$y) + runif (npts) * diff (range (verts$y))
)
edges <- match_pts_to_graph (graph, xy)
graph [edges, ] # The edges of the graph closest to `xy`
#>        geom_num edge_id    from_id from_lon from_lat      to_id   to_lon
#> 6207        203    6207 1388481924 76.38383 15.34743 1388482509 76.38860
#> 3481         89    3481 7793307898 76.43574 15.31712 2588146109 76.43894
#> 6209        203    6209 1388482509 76.38860 15.34745 1204772662 76.38895
#> 6247        203    6247 1204772780 76.40951 15.35236 1204772868 76.41809
#> 1157         35    1157 1388483314 76.45692 15.35031  977490139 76.45748
#> 3481.1       89    3481 7793307898 76.43574 15.31712 2588146109 76.43894
#> 5913        183    5913 1376768773 76.47616 15.30970  313796433 76.47604
#> 4745        133    4745 2398957576 76.47415 15.31172 1427594104 76.47425
#> 3395         89    3395 2588119056 76.42341 15.31717 2588146107 76.42349
#> 6207.1      203    6207 1388481924 76.38383 15.34743 1388482509 76.38860
#>          to_lat         d d_weighted      highway    way_id component      time
#> 6207   15.34745 511.92811 1023.85621      primary 835018468         2 368.58824
#> 3481   15.31720 343.54155  429.42694 unclassified 252786290         1 247.34992
#> 6209   15.34748  38.16886   76.33773      primary 835018468         2  27.48158
#> 6247   15.35157 925.87357 1851.74714      primary 835018468         2 666.62897
#> 1157   15.35036  61.05412  122.10824      primary  53658844         2  43.95897
#> 3481.1 15.31720 343.54155  429.42694 unclassified 252786290         1 247.34992
#> 5913   15.30953  23.13809   46.27618      primary 652570479         1  16.65942
#> 4745   15.31121  58.13286   96.88811    secondary 327102382         1  41.85566
#> 3395   15.31746  33.00845   41.26056 unclassified 252786290         1  23.76608
#> 6207.1 15.34745 511.92811 1023.85621      primary 835018468         2 368.58824
#>        time_weighted
#> 6207       737.17647
#> 3481       309.18740
#> 6209        54.96316
#> 6247      1333.25794
#> 1157        87.91793
#> 3481.1     309.18740
#> 5913        33.31885
#> 4745        69.75944
#> 3395        29.70760
#> 6207.1     737.17647